// src/forgetrustee/frontend/src/types/generated/dataTrusteeWizardRequestDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { DataTrusteeWizardBusinessRequestDto } from './dataTrusteeWizardBusinessRequestDto';
import type { DataTrusteeWizardCoreRequestDto } from './dataTrusteeWizardCoreRequestDto';
import type { DataTrusteeWizardDataRequestDto } from './dataTrusteeWizardDataRequestDto';
import type { DataTrusteeWizardImplementationRequestDto } from './dataTrusteeWizardImplementationRequestDto';
import type { DataTrusteeWizardObjectivesRequestDto } from './dataTrusteeWizardObjectivesRequestDto';

export interface DataTrusteeWizardRequestDto {
  core: DataTrusteeWizardCoreRequestDto;
  data: DataTrusteeWizardDataRequestDto;
  implementation: DataTrusteeWizardImplementationRequestDto;
  business: DataTrusteeWizardBusinessRequestDto;
  objectives: DataTrusteeWizardObjectivesRequestDto;
}