// src/forgetrustee/frontend/src/types/generated/metadataModuleIndexSummaryDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { MetadataModuleSummaryDto } from './metadataModuleSummaryDto';

export interface MetadataModuleIndexSummaryDto {
  /** List of module summaries */
  modules: MetadataModuleSummaryDto[];
}